+++ custom_components/padspan_ha/const.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
 DOMAIN = "padspan_ha"
-NAME = "PadSpan HA"
-VERSION = "0.3.7"
+NAME = "PadSpan"
+VERSION = "0.3.10"
 
 CONF_ENABLE_CLOUD = "enable_cloud"
 CONF_HUB_URL = "hub_url"